Move own service into own package
Since search lives outside of the cmd/frontend directory, this caused an invalid import from `internal` into a `cmd` package. This resolves it. Also it makes it accessible to other services in the future, like worker.
