+++ enterprise/cmd/frontend/internal/own/init.go
@@ -3,7 +3,6 @@ package own
 import (
 	"context"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/enterprise"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/own/resolvers"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel"
@@ -11,6 +10,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
+	"github.com/sourcegraph/sourcegraph/internal/own"
 )
 
 // Init initializes the given enterpriseServices to include the required
@@ -24,7 +24,7 @@ func Init(
 	enterpriseServices *enterprise.Services,
 ) error {
 	g := gitserver.NewClient()
-	o := backend.NewOwnService(g, db)
+	o := own.NewService(g, db)
 	enterpriseServices.OwnResolver = resolvers.New(db, o)
 	return nil
 }
